# perf: common utils. 包括Md5, PK, PK2, UUID, SortedTags, DictedTagstring等函数