--- a/common/utils/md5.go
+++ b/common/utils/md5.go
@@ -16,13 +16,10 @@ package utils
 
 import (
 	"crypto/md5"
-	"fmt"
-	"io"
+	"encoding/hex"
 )
 
 func Md5(raw string) string {
-	h := md5.New()
-	io.WriteString(h, raw)
-
-	return fmt.Sprintf("%x", h.Sum(nil))
+	h := md5.Sum([]byte(raw))
+	return hex.EncodeToString(h[:])
 }