--- a/common/utils/objpool.go
+++ b/common/utils/objpool.go
@@ -1,0 +1,8 @@
+package utils
+
+import (
+	"bytes"
+	"sync"
+)
+
+var bufferPool = sync.Pool{New: func() interface{} { return new(bytes.Buffer) }}